Renamed RSSPostManager to RSSPostHelpers, cleaned up
