+++ src/helpers/opmlImport.ts
@@ -1,5 +1,5 @@
 import { Feed } from '../models/Feed';
-import { RSSFeedManager } from './RSSPostManager';
+import { fetchFeedFromUrl } from './RSSPostHelpers';
 import { getHttpsUrl } from './urlUtils';
 import { Debug } from './Debug';
 import { timeout } from './Utils';
@@ -63,7 +63,7 @@ const convertOPMLFeed = async (opmlFeed: OPMLFeed): Promise<Feed | undefined> =>
     const feedUrl = getHttpsUrl(opmlFeed.feedUrl);
     Debug.log('convertOPMLFeed', {feedUrl});
     try {
-        const feed = await timeout(15000, RSSFeedManager.fetchFeedFromUrl(feedUrl));
+        const feed = await timeout(15000, fetchFeedFromUrl(feedUrl));
         Debug.log('convertOPMLFeed', 'after timeout', {feed});
         if (feed == null) {
             return undefined;
