+++ test/helpers/RSSPostManagerTest.ts
@@ -1,11 +1,11 @@
-import { RSSPostManager, RSSFeedManager } from '../../src/helpers/RSSPostManager';
+import { htmlToMarkdown, extractTextAndImagesFromMarkdown, getFeedFromHtml, isTitleSameAsText } from '../../src/helpers/RSSPostHelpers';
 
 test('Parse CDATA descriptions from RSS', async () => {
     const text = 'text';
     const description = `<![CDATA[${text}]]>`;
     const expectedResult = text;
 
-    const result = RSSPostManager.htmlToMarkdown(description);
+    const result = htmlToMarkdown(description);
 
     expect(result).toBe(expectedResult);
 });
@@ -15,7 +15,7 @@ test('Parse image from RSS', async () => {
     const description = `<img src="${link}" alt="">`;
     const expectedResult = `![](${link})`;
 
-    const result = RSSPostManager.htmlToMarkdown(description);
+    const result = htmlToMarkdown(description);
 
     expect(result).toBe(expectedResult);
 });
@@ -25,7 +25,7 @@ test('Parse CDATA descriptions containing an image from RSS', async () => {
     const description = `<![CDATA[<img src="${link}" alt="">]]>`;
     const expectedResult = `![](${link})`;
 
-    const result = RSSPostManager.htmlToMarkdown(description);
+    const result = htmlToMarkdown(description);
 
     expect(result).toBe(expectedResult);
 });
@@ -35,7 +35,7 @@ test('Parse html with image', () => {
     const description = `<div><img src="${link}" alt=""></div>`;
     const expectedResult = `![](${link})`;
 
-    const result = RSSPostManager.htmlToMarkdown(description);
+    const result = htmlToMarkdown(description);
 
     expect(result).toBe(expectedResult);
 });
@@ -45,7 +45,7 @@ test('Parse image from markdown', () => {
     const markdown = `![](${link})`;
     const expectedResult = ['', [{uri: link}]];
 
-    const result = RSSPostManager.extractTextAndImagesFromMarkdown(markdown, '');
+    const result = extractTextAndImagesFromMarkdown(markdown, '');
 
     expect(result).toEqual(expectedResult);
 });
@@ -56,7 +56,7 @@ test('Parse link', () => {
     const description = `<a href="${link}">description</a>`;
     const expectedResult = `[description](${link})`;
 
-    const result = RSSPostManager.htmlToMarkdown(description);
+    const result = htmlToMarkdown(description);
 
     expect(result).toBe(expectedResult);
 });
@@ -69,7 +69,7 @@ test('Parse description with multiline CDATA', () => {
     `;
     const expectedResult = 'a';
 
-    const result = RSSPostManager.htmlToMarkdown(description);
+    const result = htmlToMarkdown(description);
 
     expect(result).toBe(expectedResult);
 });
@@ -94,7 +94,7 @@ test('Fetch RSS feed from URL', async () => {
     const expectedFavicon = 'https://index.hu/assets/images/favicons/apple-touch-icon.png';
     const expectedName = 'Index';
 
-    const result = await RSSFeedManager.getFeedFromHtml(baseUrl, htmlContent);
+    const result = await getFeedFromHtml(baseUrl, htmlContent);
 
     expect(result).not.toBeNull();
 
@@ -109,7 +109,7 @@ test('Fetch RSS feed from URL', async () => {
 test('Compare text with title with links removed', () => {
     const title = 'Monvid is based on Blockchain technology and is a decentralized application which relays on streaming nodes provided by the community. Everyone can join this community and share their resources with the platform.↵#ICO #MVID #Monvid';
     const text = 'Monvid is based on Blockchain technology and is a decentralized application which relays on streaming nodes provided by the community. Everyone can join this community and share their resources with the platform.[#ICO](https://twitter.com/hashtag/ICO?src=hash) [#MVID](https://twitter.com/hashtag/MVID?src=hash) [#Monvid](https://twitter.com/hashtag/Monvid?src=hash)';
-    const result = RSSPostManager.isTitleSameAsText(title, text);
+    const result = isTitleSameAsText(title, text);
 
     expect(result).toBeTruthy();
 });
